refactor(agent): extract prompt rendering into agentprompt package
Move the agent's bootstrap prompt rendering out of driver.go into a new
internal/agent/agentprompt package: the //go:embed PROMPT.md + promptText,
promptTemplate with the RenderEvent FuncMap, RenderEvent, PROMPT.md itself,
and the golden tests.

The package exposes Render(started, prompt, events) taking its inputs as
parameters, so it depends only on the proto package and not on
internal/agent (no import cycle). driver.go's Config.prompt is replaced by
a direct call to agentprompt.Render(cfg.Started, cfg.Prompt, events).

Pure, behavior-preserving refactor: the golden files are unchanged (moved
verbatim) and the rendered output is byte-for-byte identical.
