+++ internal/agent/agentprompt/testdata/prompt-first-run.golden
