+++ internal/agent/agentprompt/testdata/prompt-wake-empty.golden
