+++ internal/agent/agentprompt/testdata/prompt-wake-events-workspace.golden
