+++ internal/agent/agentprompt/testdata/prompt-wake-events.golden
