+++ internal/agent/driver_test.go
@@ -8,15 +8,12 @@ import (
 	"strings"
 	"syscall"
 	"testing"
-	"time"
 
 	xagentv1 "github.com/icholy/xagent/internal/proto/xagent/v1"
 	"github.com/icholy/xagent/internal/xagentclient"
 	"google.golang.org/protobuf/testing/protocmp"
-	"google.golang.org/protobuf/types/known/timestamppb"
 	"gotest.tools/v3/assert"
 	"gotest.tools/v3/assert/cmp"
-	"gotest.tools/v3/golden"
 )
 
 // testTaskVersion is the version returned by setupDriver's GetTask mock. The
@@ -406,68 +403,4 @@ func TestDriverRun_SetupCommandOutputTeed(t *testing.T) {
 	assert.NilError(t, err)
 	assert.Assert(t, cmp.Contains(string(got), "out-marker"))
 	assert.Assert(t, cmp.Contains(string(got), "err-marker"))
-}
-
-// TestConfigPromptGolden snapshots the whole rendered bootstrap prompt across
-// its branches: the first-run get_my_task bootstrap, a wake that injects the
-// pending events as a JSON array, the bare fallback when a wake has nothing
-// pending, and a wake with a workspace prompt appended.
-// Regenerate the goldens with: go test ./internal/agent/ -run TestConfigPromptGolden -update
-func TestConfigPromptGolden(t *testing.T) {
-	t.Parallel()
-	// Fixed timestamps keep the rendered createdAt fields stable across runs.
-	events := []*xagentv1.Event{
-		{
-			Id:        42,
-			CreatedAt: timestamppb.New(time.Unix(1_700_000_000, 0).UTC()),
-			Payload: &xagentv1.Event_External{External: &xagentv1.ExternalPayload{
-				Description: "PR review requested",
-				Url:         "https://github.com/icholy/xagent/pull/1394",
-			}},
-		},
-		{
-			Id:        43,
-			CreatedAt: timestamppb.New(time.Unix(1_700_000_100, 0).UTC()),
-			Payload: &xagentv1.Event_Instruction{Instruction: &xagentv1.InstructionPayload{
-				Text: "keep going",
-				Url:  "https://github.com/icholy/xagent/issues/2",
-			}},
-		},
-	}
-	tests := []struct {
-		name   string
-		cfg    *Config
-		events []*xagentv1.Event
-		golden string
-	}{
-		{
-			name:   "first run bootstraps via get_my_task",
-			cfg:    &Config{},
-			golden: "prompt-first-run.golden",
-		},
-		{
-			name:   "wake injects pending events",
-			cfg:    &Config{Started: true},
-			events: events,
-			golden: "prompt-wake-events.golden",
-		},
-		{
-			name:   "wake with nothing pending falls back",
-			cfg:    &Config{Started: true},
-			golden: "prompt-wake-empty.golden",
-		},
-		{
-			name:   "wake injects events with a workspace prompt appended",
-			cfg:    &Config{Started: true, Prompt: "Custom workspace instructions."},
-			events: events,
-			golden: "prompt-wake-events-workspace.golden",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := tt.cfg.prompt(tt.events)
-			assert.NilError(t, err)
-			golden.Assert(t, got, tt.golden)
-		})
-	}
 }
